Built out tutorial and enemy attack text
Started building a text based system for tutorializing the game and adding text whenever the enemy attacks.  2 known issues, 1 the tutorial dialogue box has to be placed oddly in the scene to get it where it needs to be in game view.  Need to figure out why.  2 something is destroying the enemy text gameobject before text is added to it.